Returning to Int, because Short created strange errors in Serializer

+++ app/src/commonMain/kotlin/model/Geyser.kt
@@ -36,13 +36,13 @@ data class Geyser(
     val emitRate: Int,
 
     /** Average emit rate in gram per second. */
-    val avgEmitRate: Short,
+    val avgEmitRate: Int,
 
     /** Idle time after eruption in seconds. */
-    val idleTime: Short,
+    val idleTime: Int,
 
     /** Duration of eruption in seconds. */
-    val eruptionTime: Short,
+    val eruptionTime: Int,
 
     /** Count of dormancy cycles. */
     val dormancyCycles: Float,
